Add test settings to allow workers environments to be reused
Reviewed-on: http://review.couchbase.org/43423
Tested-by: buildbot <build@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>

+++ perfrunner/helpers/worker.py
@@ -10,7 +10,6 @@ from spring.wgen import WorkloadGen
 from sqlalchemy import create_engine
 
 from perfrunner import celerylocal, celeryremote
-from perfrunner.helpers.misc import uhex
 from perfrunner.settings import REPO
 
 
@@ -45,7 +44,8 @@ class RemoteWorkerManager(object):
         self.cluster_spec = cluster_spec
         self.buckets = test_config.buckets or test_config.max_buckets
 
-        self.temp_dir = '/tmp/{}'.format(uhex()[:12])
+        self.reuse_worker = test_config.worker_settings.reuse_worker
+        self.temp_dir = test_config.worker_settings.worker_dir
         self.user, self.password = cluster_spec.client_credentials
         with settings(user=self.user, password=self.password):
             self.initialize_project()
@@ -61,6 +61,14 @@ class RemoteWorkerManager(object):
 
                 qname = '{}-{}'.format(master.split(':')[0], bucket)
                 temp_dir = '{}-{}'.format(self.temp_dir, qname)
+
+                r = run('test -d {}'.format(temp_dir), warn_only=True, quiet=True)
+                if r.return_code == 0:
+                    if self.reuse_worker == 'true':
+                        return
+                    logger.error('Worker env exists, but reuse not specified')
+                    sys.exit(1)
+
                 run('mkdir {}'.format(temp_dir))
                 with cd(temp_dir):
                     run('git clone {}'.format(REPO))
@@ -114,7 +122,8 @@ class RemoteWorkerManager(object):
                     logger.info('Cleaning up remote worker environment')
                     qname = '{}-{}'.format(master.split(':')[0], bucket)
                     temp_dir = '{}-{}'.format(self.temp_dir, qname)
-                    run('rm -fr {}'.format(temp_dir))
+                    if self.reuse_worker == 'false':
+                        run('rm -fr {}'.format(temp_dir))
 
 
 class LocalWorkerManager(RemoteWorkerManager):
